Convert OidcClient to TypeScript.
