+++ package.json
@@ -22,6 +22,7 @@
     "gulp": "^3.8.11",
     "gulp-concat": "^2.5.2",
     "gulp-rename": "^1.2.0",
+    "gulp-typescript": "^2.10.0",
     "gulp-uglify": "^1.1.0"
   }
 }
